Add JSON file with information about Git releases
The “git-versions.json” file contains information about the public
releases of Git. This includes all stable versions along with their
release dates as well as a list of all publicly disclosed Git
vulnerabilities and which versions they affect.

The intent of this file is to be able to discern which Git versions are
not recommended for productive use, because they are outdated or
affected by known vulnerabilities.

This file is up-to-date as of 2019-02-14, including CVE-2018-17456.

Git versions before 2.0.0 and vulnerabilities in versions released
before 2017 were omitted for simplicity, because all Git versions
earlier than that are affected by at least one vulnerability known
before this was implemented anyway.