Add semantic versioning utilities
These functions add support for parsing and comparing versions according
to the semantic versioning specification [1]. In addition, a function is
provided to check whether a given version fulfills a compound version
requirement such as “>2.9 <= 2.19.2”.

All functions are covered by unit tests. This temporarily allows the
functions related to semantic versioning to be unused in the ESLint
configuration, because they will only be used in a subsequent commit,
and ESLint would complain otherwise.

[1] https://semver.org/