Implement area chart
This adds area charts as a new chart type. Area charts visualize
datasets not as single lines starting from 0 but as stacked areas. Area
charts are useful to visualize data that is comprised of multiple parts
that add up to a total. In this way, the sum of all datasets can easily
be tracked as well as the relative distribution of the individual parts.